A track needs init if the first frame has not been read.
The weird logic of to_decode and fresh misfired when having decoded the first frame already; and the "fresh" part does not apply because we did actually start in the middle of the stream (so no Lame tag as clear marker of change).

Questions remain: Why don't I always return a header change on first frame? Is this "fresh" frame property really useful?