Reworked common module loader code with platform support from libcompat.
The ugliness of various wrappery for working out module paths is now
inside src/compat/compat.c . The module loading as such is not the issue,
but the construction of proper path arguments is, to ensure that
Unicode and possibly long paths work properly in Windows environments.

This is the work of JonY from the first loader_rework branch rolled
into compat.c, mostly. The code for Windows is not obviously broken.
I tested that it works using MinGW under wine inside a 32 bit chroot
(no kidding). I got it crashing when enabling debugging, but that can
have various reasons in the chain.

Some fixes emerged together with the intended changes, like updating
the legacy module shim for bindir and extending the compat file opening
to use the full deal of long Unicode paths (real testing needs to be
done). I am a bit confused by the needed changes to congfiure (like
explicitly adding -lshlwapi). Are those not necessary with MinGW-w64
on native Windows?