out123: fix output buffer allocation for resampling, add --speed
The mono resampler seems to work now. I can generate a wave pattern and
speed it up/down. Still need to fix the situtation with dithering.

And test the dithering.

Generally, I need to write some tests for the various format conversion
and mixing options. I can synthetize test data on the fly, or use some
example recording to take the syn123 generators out of the picture.